Add snapshot testing - Update operate.test.js and calculate.test.js with new paths. - Add new directories > unit-testing and snapshot-testing and move all relevant files. - Install react-test-renderer.
