+++ src/__test__/unit-testing/calculate.test.js
@@ -1,4 +1,4 @@
-import calculate from '../logic/calculate';
+import calculate from '../../logic/calculate';
 
 describe('testing the calculate function ', () => {
   const obj = { total: null, next: null, operation: null };
